// frontend/src/features/data_entry/stores/types.ts
export type DataEntryDispatch = Dispatch<DataEntryAction>;

export type DataEntryAction =
  | {
      type: "DATA_ENTRY_LOADED";
      dataEntry: GetDataEntryResponse;
    }
  | {
      type: "DATA_ENTRY_LOAD_FAILED";
      error: AnyApiError;
    }
  | {
      type: "DATA_ENTRY_SAVE_FAILED";
      error: AnyApiError;
    }
  | {
      type: "DATA_ENTRY_NOT_FOUND";
    }
  | {
      type: "FORM_SAVE_FAILED";
      error: AnyApiError;
    }
  | {
      type: "FORM_SAVED";
      data: PollingStationResults;
      validationResults: ValidationResults;
      aborting: boolean;
      continueToNextSection: boolean;
    }
  | {
      type: "SET_STATUS";
      status: Status;
    }
  | {
      type: "SET_CACHE";
      cache: TemporaryCache;
    }
  | {
      type: "UPDATE_FORM_SECTION";
      partialFormSection: Partial<FormSection>;
    }
  | {
      type: "RESET_TARGET_FORM_SECTION";
    }
  | {
      type: "REGISTER_CURRENT_FORM";
      form: FormSectionReference;
    };
// ...
export type FormSectionReference =
  | {
      id: "recounted";
      type: "recounted";
    }
  | {
      id: "voters_votes_counts";
      type: "voters_and_votes";
    }
  | {
      id: "differences_counts";
      type: "differences";
    }
  | {
      id: `political_group_votes_${number}`;
      type: "political_group_votes";
      number: number;
    }
  | {
      id: "save";
      type: "save";
    };

//store unvalidated data
export type TemporaryCache = {
  key: FormSectionId;
  data: Partial<PollingStationResults>;
};

// Status of the form controller
// This is a type instead of an enum because of https://github.com/ArnaudBarre/eslint-plugin-react-refresh/issues/36
export type Status = "idle" | "saving" | "deleting" | "deleted" | "finalising" | "finalised" | "aborted";
// ...
export type FormSection = {
  index: number; //fixate the order of filling in sections
  id: FormSectionId;
  title?: string;
  hasChanges: boolean;
  isSaved: boolean; //whether this section has been sent to the server
  isSubmitted?: boolean; //whether this section has been submitted in the latest request
  acceptWarnings: boolean;
  acceptWarningsError: boolean;
  errors: ValidationResult[];
  warnings: ValidationResult[];
};